requirements.txt
Flask==2.1.3
Flask_Admin==1.6.0
Flask_Login==0.6.1
Flask_Mail==0.9.1
Flask_SQLAlchemy==2.5.1
Flask_WTF==1.0.1
Pillow==10.0.0
SQLAlchemy==1.4.39
Werkzeug==2.1.2
WTForms==3.0.1
